fixed compilation issues
