+++ src/ts/element.ts
@@ -2,7 +2,6 @@ import type { Writable } from 'svelte/store';
 import type { Models } from '../types/models';
 import type { Camera } from './camera';
 import type Svelte from '../svelte/Main.svelte';
-import type { MicrioUIProps } from '../svelte/Main.svelte';
 
 import { once, deepCopy, fetchJson, jsonCache, fetchInfo, fetchAlbumInfo, idIsV5 } from './utils';
 import { ATTRIBUTE_OPTIONS as AO, BASEPATH, BASEPATH_V5, localStorageKeys } from './globals';
@@ -123,7 +122,7 @@ export class HTMLMicrioElement extends HTMLElement {
 	/** The Svelte UI component instance.
 	 * @internal
 	*/
-	_ui:{setProps:(p:Partial<MicrioUIProps>) => void}|undefined;
+	_ui:{setProps?:(p:Partial<MicrioUIProps>) => void}|undefined;
 
 	/** Custom settings object provided programmatically, overriding server-fetched settings. */
 	public defaultSettings?:Partial<Models.ImageInfo.Settings> = this.defaultSettings;
@@ -279,7 +278,7 @@ export class HTMLMicrioElement extends HTMLElement {
 	private loadArchiveBin(path:string, id:string) : Promise<void> {
 		this.printUI(true, true); // Print minimal UI for loading progress
 		// Use the archive utility to load and report progress
-		return archive.load(path, 'g/'+id, loadingProgress => this._ui?.setProps({loadingProgress}));
+		return archive.load(path, 'g/'+id, loadingProgress => this._ui?.setProps?.({loadingProgress}));
 	}
 
 	/**
@@ -371,8 +370,9 @@ export class HTMLMicrioElement extends HTMLElement {
 	 * @param noLogo If true, hides the Micrio logo.
 	 */
 	private printUI(noHTML:boolean, noLogo:boolean) : void {
+		//@ts-ignore
 		if(!this._ui) this._ui = mount(HTMLMicrioElement.Svelte, {target:this, props:{micrio:this,noHTML,noLogo}})
-		else this._ui.setProps({noHTML, noLogo});
+		else this._ui.setProps?.({noHTML, noLogo});
 	}
 
 	/**
@@ -381,7 +381,7 @@ export class HTMLMicrioElement extends HTMLElement {
 	 * @param str The error message string.
 	 */
 	printError(str?:string) : void {
-		this._ui?.setProps({error: str??'An unknown error has occurred'});
+		this._ui?.setProps?.({error: str??'An unknown error has occurred'});
 	}
 
 	/**
@@ -730,4 +730,18 @@ export class HTMLMicrioElement extends HTMLElement {
 	get lang() { return get(this._lang) }
 	/** Setter for the current language code. Triggers language change logic. */
 	set lang(l:string) { this.setAttribute('lang', l) }
+}
+
+// --- Svelte UI Props ---
+export interface MicrioUIProps {
+	/** The main HTMLMicrioElement instance. Provided by element.ts */
+	micrio: HTMLMicrioElement;
+	/** If true, suppresses rendering of most UI elements (except markers if data-ui="markers"). */
+	noHTML: boolean;
+	/** If true, suppresses rendering of the Micrio logo. Defaults to `noHTML`. */
+	noLogo?: boolean;
+	/** Loading progress (0-1), used for the progress indicator. */
+	loadingProgress?: number;
+	/** Optional error message to display. */
+	error?: string|undefined;
 }
